requirements.txt
Django==2.0.2
geoip2==2.7.0
django-daterange-filter==1.3.0
django_jinja==2.4.1
